# common/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.0.0-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers (schema and file descriptor)\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `http://MY-REGISTRY-URL/api` by default. Therefore you must prefix all API operation paths with `/api` in this case. For example: `/api/ids/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
        "license": {
...
        }
...
    "paths": {
        "/rules": {
            "summary": "Manage the global rules that apply to all artifacts if not otherwise configured.",
            "get": {
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RuleType"
                                    }
                                }
                            }
                        },
                        "description": "The list of names of the globally configured rules."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "listGlobalRules",
                "summary": "List global rules",
                "description": "Gets a list of all the currently configured global rules (if any).\n\nThis operation can fail for the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rule"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "204": {
                        "description": "The global rule was added."
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "createGlobalRule",
                "summary": "Create global rule",
                "description": "Adds a rule to the list of globally configured rules.\n\nThis operation can fail for the following reasons:\n\n* The rule type is unknown (HTTP error `400`)\n* The rule already exists (HTTP error `409`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "204": {
                        "description": "All global rules have been removed successfully."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteAllGlobalRules",
                "summary": "Delete all global rules",
                "description": "Deletes all globally configured rules.\n\nThis operation can fail for the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            }
        },
        "/ids/{globalId}": {
            "summary": "Access artifact content utilizing an artifact version's globally unique identifier.",
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "OpenAPI Example": {
                                        "value": {
                                            "openapi": "3.0.2",
                                            "info": {
                                                "title": "Empty API",
                                                "version": "1.0.0",
                                                "description": "An example API design using OpenAPI."
                                            }
                                        }
                                    }
                                }
                            },
                            "application/x-protobuf": {
                                "examples": {
                                    "Protobuf Example": {
                                        "value": "message SearchResponse {\n  repeated Result result = 1;\n}\n\nmessage Result {\n  required string url = 1;\n  optional string title = 2;\n  repeated string snippets = 3;\n}"
                                    }
                                }
                            },
                            "application/x-protobuffer": {
                                "examples": {
                                    "Protobuffer Example": {
                                        "value": "message SearchResponse {\n  repeated Result result = 1;\n}\n\nmessage Result {\n  required string url = 1;\n  optional string title = 2;\n  repeated string snippets = 3;\n}"
                                    }
                                }
                            }
                        },
                        "description": "The content of the artifact version."
                    }
                },
                "operationId": "getArtifactByGlobalId",
                "summary": "Get artifact by global ID",
                "description": "Gets the content for an artifact version in the registry using its globally unique\nidentifier.\n\nThis operation may fail for one of the following reasons:\n\n* No artifact version with this `globalId` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "globalId",
                    "description": "Global identifier for an artifact version.",
                    "schema": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/ids/{globalId}/meta": {
            "summary": "Access an artifact verison's metadata using its globally unique identifier.",
            "get": {
                "tags": [
                    "Metadata"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArtifactMetaData"
                                }
                            }
                        },
                        "description": "The artifact's metadata."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getArtifactMetaDataByGlobalId",
                "summary": "Get global artifact metadata",
                "description": "Gets the metadata for an artifact version in the registry using its globally unique\nidentifier.  The returned metadata includes both generated (read-only) and editable\nmetadata (such as name and description).\n\nThis operation may fail for one of the following reasons:\n\n* No artifact version with this `globalId` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "globalId",
                    "description": "Global identifier for an artifact version.",
                    "schema": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/artifactgroups/{groupId}/artifacts/{artifactId}/rules/{rule}": {
            "summary": "Manage the configuration of a single artifact rule.",
            "get": {
                "tags": [
                    "Artifact rules"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "Information about a rule."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getArtifactRuleConfig",
                "summary": "Get artifact rule configuration",
                "description": "Returns information about a single rule configured for an artifact.  This is useful\nwhen you want to know what the current configuration settings are for a specific rule.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No rule with this name/type is configured for this artifact (HTTP error `404`)\n* Invalid rule type (HTTP error `400`)\n* A server error occurred (HTTP error `500`)"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rule"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Artifact rules"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "Rule configuration was updated."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateArtifactRuleConfig",
                "summary": "Update artifact rule configuration",
                "description": "Updates the configuration of a single rule for the artifact.  The configuration data\nis specific to each rule type, so the configuration of the `COMPATIBILITY` rule \nis in a different format from the configuration of the `VALIDITY` rule.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No rule with this name/type is configured for this artifact (HTTP error `404`)\n* Invalid rule type (HTTP error `400`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Artifact rules"
                ],
                "responses": {
                    "204": {
                        "description": "The rule was successfully deleted."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteArtifactRule",
                "summary": "Delete artifact rule",
                "description": "Deletes a rule from the artifact.  This results in the rule no longer applying for\nthis artifact.  If this is the only rule configured for the artifact, this is the \nsame as deleting **all** rules, and the globally configured rules now apply to\nthis artifact.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No rule with this name/type is configured for this artifact (HTTP error `404`)\n* Invalid rule type (HTTP error `400`)\n* A server error occurred (HTTP error `500`)"
            },
            "parameters": [
                {
                    "name": "groupId",
                    "description": "The artifact group ID.  Must be a string provided by the client - represents the name of the grouping of artifacts.",
                    "schema": {
                        "$ref": "#/components/schemas/GroupId"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "artifactId",
                    "description": "The artifact ID.  Can be a string (client-provided) or UUID (server-generated) representing the unique artifact identifier.",
                    "schema": {
                        "$ref": "#/components/schemas/ArtifactId"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "rule",
                    "description": "The unique name/type of a rule.",
                    "schema": {
                        "enum": [
                            "VALIDITY",
                            "COMPATIBILITY"
                        ],
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/rules/{rule}": {
            "summary": "Manage the configuration of a single global artifact rule.",
            "get": {
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "The global rule's configuration."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getGlobalRuleConfig",
                "summary": "Get global rule configuration",
                "description": "Returns information about the named globally configured rule.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rule"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "The global rule's configuration was successfully updated."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateGlobalRuleConfig",
                "summary": "Update global rule configuration",
                "description": "Updates the configuration for a globally configured rule.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Global rules"
                ],
                "responses": {
                    "204": {
                        "description": "The global rule was successfully deleted."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteGlobalRule",
                "summary": "Delete global rule",
                "description": "Deletes a single global rule.  If this is the only rule configured, this is the same\nas deleting **all** rules.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* Rule cannot be deleted (HTTP error `409`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "rule",
                    "description": "The unique name/type of a rule.",
                    "schema": {
                        "$ref": "#/components/schemas/RuleType"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "Rule": {
                "title": "Root Type for Rule",
                "description": "",
                "required": [
                    "config"
                ],
                "type": "object",
                "properties": {
                    "config": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/RuleType"
                    }
                },
                "example": {
                    "type": "VALIDITY",
                    "config": "FULL"
                }
            },
            "Error": {
                "title": "Root Type for Error",
                "description": "All error responses, whether `4xx` or `5xx` will include one of these as the response\nbody.",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "The short error message.",
                        "type": "string"
                    },
                    "error_code": {
                        "format": "int32",
                        "description": "The server-side error code.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "Full details about the error.  This might contain a server stack trace, for example.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The error name - typically the classname of the exception thrown by the server.",
                        "type": "string"
                    }
                },
                "example": {
                    "error_code": 500,
                    "message": "An error occurred somewhere."
                }
            },
            "RuleType": {
                "description": "",
                "enum": [
                    "VALIDITY",
                    "COMPATIBILITY"
                ],
                "type": "string",
                "example": "VALIDITY",
                "x-codegen-package": "io.apicurio.registry.types"
            },
            "ArtifactType": {
                "description": "",
                "enum": [
                    "AVRO",
                    "PROTOBUF",
                    "PROTOBUF_FD",
                    "JSON",
                    "OPENAPI",
                    "ASYNCAPI",
                    "GRAPHQL",
                    "KCONNECT",
                    "WSDL",
                    "XSD",
                    "XML"
                ],
                "type": "string",
                "example": "AVRO",
                "x-codegen-package": "io.apicurio.registry.types"
            },
...
            "ArtifactState": {
                "description": "Describes the state of an artifact or artifact version.  The following states\nare possible:\n\n* ENABLED\n* DISABLED\n* DEPRECATED\n",
                "enum": [
                    "ENABLED",
                    "DISABLED",
                    "DEPRECATED"
                ],
                "type": "string",
                "x-codegen-package": "io.apicurio.registry.types"
            },
...
            "Properties": {
                "description": "User-defined name-value pairs. Name and value must be strings.",
                "type": "object",
                "additionalProperties": {
                    "type": "string"
                }
            },
...
            "GroupId": {
                "description": "An id of a single Artifact Group.",
                "type": "string",
                "example": "\"my-group\""
            },
            "ArtifactId": {
                "description": "The ID of a single Artifact.",
                "type": "string",
                "example": "\"example-artifact\""
            },
...
            "ArtifactMetaData": {
                "title": "Root Type for ArtifactMetaData",
                "description": "",
                "required": [
                    "id",
                    "createdBy",
                    "createdOn",
                    "modifiedBy",
                    "modifiedOn",
                    "version",
                    "type",
                    "globalId",
                    "state",
                    "group"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "createdOn": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "modifiedBy": {
                        "type": "string"
                    },
                    "modifiedOn": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "id": {
                        "$ref": "#/components/schemas/ArtifactId",
                        "description": ""
                    },
                    "version": {
                        "format": "int64",
                        "description": "",
                        "type": "integer"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ArtifactType",
                        "description": ""
                    },
                    "globalId": {
                        "format": "int64",
                        "description": "",
                        "type": "integer"
                    },
                    "state": {
                        "$ref": "#/components/schemas/ArtifactState",
                        "description": ""
                    },
                    "labels": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "properties": {
                        "$ref": "#/components/schemas/Properties",
                        "description": ""
                    },
                    "groupId": {
                        "$ref": "#/components/schemas/GroupId",
                        "description": ""
                    }
                },
                "example": {
                    "group": "My-Group",
                    "id": "Procurement-Invoice",
                    "name": "Artifact Name",
                    "description": "Description of the artifact",
                    "type": "AVRO",
                    "version": 18,
                    "createdBy": "user1",
                    "createdOn": "2019-03-22T12:51:19Z",
                    "modifiedBy": "user2",
                    "modifiedOn": "2019-07-19T15:09:00Z",
                    "globalId": 12984719247,
                    "labels": [
                        "label-1",
                        "label-2"
                    ],
                    "properties": {
                        "custom-1": "foo",
                        "custom-2": "bar"
                    }
                }
            }
        },
        "responses": {
            "NotFound": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "NotFoundExample": {
                                "value": {
                                    "error_code": 404,
                                    "message": "No artifact with id 'Topic-1/Inbound' could be found."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can return a `404` error."
            },
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
            "BadRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                },
                "description": "Common response for all operations that can return a `400` error."
            },
            "Conflict": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ConflictExample": {
                                "value": {
                                    "error_code": 409,
                                    "message": "The artifact content was invalid."
                                }
                            }
                        }
                    }
                },
                "description": "Common response used when an input conflicts with existing data."
            },
...
        }
    },
...
}